refactor(update): extract config hot-reload handlers to update/config.rs (commit 5/N)
Pull six hot-reload-related Message variants out of mod.rs into a new
update/config.rs file with one handler method per variant:

- Message::ToggleLightMode → handle_toggle_light_mode (17 lines)
- Message::VisualizerConfigChanged → handle_visualizer_config_changed (17)
- Message::ThemeConfigReloaded → handle_theme_config_reloaded (20)
- Message::SettingsConfigReloaded → handle_settings_config_reloaded (30)
- Message::SettingsReloadDataLoaded → handle_settings_reload_data_loaded (10)
- Message::HotkeyConfigUpdated → handle_hotkey_config_updated (5)

Total ~100 inline lines moved out of the central match. The mod.rs arms
collapse to thin one-line delegations. The grouping is "things that
happen when configuration mutates" — user-driven toggle, file-watcher
hot-reload, and the resulting cascade of UI/state updates.

cargo build / clippy --all-targets -D warnings / nightly fmt --check /
test --workspace all clean (662 + 368 + 1 doctest).
